factor out Digested into separate module
