+++ rtx/src/converter.rs
@@ -2,9 +2,10 @@ use rtx_core::common::error::*;
 use rtx_core::common::object::Object;
 use rtx_core::common::{Config, DataSize, OutputFormat};
 use rtx_core::document::Document;
+use rtx_core::digested::Digested;
 use rtx_core::list::List;
 use rtx_core::state::State;
-use rtx_core::{generate_message, s, Core, CoreOptions, Digested, Error, Info};
+use rtx_core::{generate_message, s, Core, CoreOptions, Error, Info};
 use rtx_package::package;
 use std::sync::Arc;
 // use std::sync::RwLockWriteGuard;
