+++ rtx_package/src/package.rs
@@ -55,7 +55,8 @@ pub use rtx_core::tokens::Tokens;
 pub use rtx_core::util::pathname;
 pub use rtx_core::util::radix;
 pub use rtx_core::whatsit::Whatsit;
-pub use rtx_core::{BoxOps, Core, Digested, DigestedData, TexMode};
+pub use rtx_core::digested::{Digested, DigestedData};
+pub use rtx_core::{BoxOps, Core, TexMode};
 
 // ------------------------------------------------------------------------------------------------
 // ------------------------------------------------------------------------------------------------
